remove the Find function
Most of the time users will use Is, As, or AsType
The Find function is easy to write if it is needed.
