+++ go_123_test.go
@@ -2,7 +2,11 @@
 
 package errors
 
-import "testing"
+import (
+	"errors"
+	"io"
+	"testing"
+)
 
 func testFindErr(err error, v int) bool {
 	for err := range UnwrapGroups(err) {
@@ -42,5 +46,49 @@ func TestUnwrapGroups(t *testing.T) {
 
 	if testFindErr(err, 32) {
 		t.Errorf("found not exists")
+	}
+}
+
+func find(origErr error, test func(error) bool) error {
+	var foundErr error
+	for err := range UnwrapGroups(origErr) {
+		if test(err) {
+			foundErr = err
+			break
+		}
+	}
+	return foundErr
+}
+
+func TestFind(t *testing.T) {
+	eNew := errors.New("error")
+	wrapped := Wrap(nilError{}, "nil")
+	tests := []struct {
+		err    error
+		finder func(error) bool
+		found  error
+	}{
+		{io.EOF, func(_ error) bool { return true }, io.EOF},
+		{io.EOF, func(_ error) bool { return false }, nil},
+		{io.EOF, func(err error) bool { return err == io.EOF }, io.EOF},
+		{io.EOF, func(err error) bool { return err != io.EOF }, nil},
+
+		{eNew, func(err error) bool { return true }, eNew},
+		{eNew, func(err error) bool { return false }, nil},
+
+		{nilError{}, func(err error) bool { return true }, nilError{}},
+		{nilError{}, func(err error) bool { return false }, nil},
+		{nilError{}, func(err error) bool { _, ok := err.(nilError); return ok }, nilError{}},
+
+		{wrapped, func(err error) bool { return true }, wrapped},
+		{wrapped, func(err error) bool { return false }, nil},
+		{wrapped, func(err error) bool { _, ok := err.(nilError); return ok }, nilError{}},
+	}
+
+	for _, tt := range tests {
+		got := find(tt.err, tt.finder)
+		if got != tt.found {
+			t.Errorf("WrapNoStack(%v): got: %q, want %q", tt.err, got, tt.found)
+		}
 	}
 }
